testと手直し、view2のソースコードに寄せる対応
