+++ Sources/YamlParserKit/YamlParser.swift
@@ -5,11 +5,9 @@ import Yaml
 public class YamlParser {
     public var views: [View] { _views }
     public var settings: Settings { _settings }
-    public var views2: [View] { _views2 }
     
     private var _views: [View] = []
     private var _settings: Settings = .init()
-    private var _views2: [View] = []
     
     public init() {}
     
@@ -30,12 +28,8 @@ public class YamlParser {
             
             let dictionary = dictionaries[startIndex]
             
-            if dictionary.key == .string("views"), case .array(let viewsArray) = dictionary.value {
-                _views = ViewsGenerator.generate(index: 0, viewsArray: viewsArray)
-            }
-            
             if dictionary.key == .string("views2"), case .dictionary(let viewsDicList) = dictionary.value {
-                _views2 = Views2Generator.generate2(viewsDicList: viewsDicList)
+                _views = ViewsGenerator.generate(viewsDicList: viewsDicList)
             }
             
             if dictionary.key == .string("settings"), case .dictionary(let settingsDictionary) = dictionary.value {
